cybercore/src/cybercore-main/requirements.txt
# Core CLI dependencies
InquirerPy>=0.3.4
pyfiglet>=0.8.post1

# Database
psycopg2-binary>=2.9.9
bcrypt>=4.1.2

# API and networking
requests>=2.31.0

# Ansible automation
ansible-runner>=2.3.4
PyYAML>=6.0.1

# Configuration
python-dotenv>=1.0.0

# Logging and utilities
colorlog>=6.8.0